Create knapsack_code.cpp
A knapsack is a type of bag that is carried on the back or over the shoulder, used for carrying supplies or personal belongings. A knapsack problem is an optimization problem where you have to choose which items to include in the knapsack so that the total value is maximized and the total weight is within a given limit. There are different types of knapsack problems, such as 0/1 knapsack problem and fractional knapsack problem5