Create matrix_chain_multiplication.c
Matrix chain multiplication is an optimization problem that aims to find the most efficient way to multiply a given sequence of matrices. The problem is not to perform the multiplications, but to decide the order of the matrix multiplications involved. The order of matrix multiplications affects the number of scalar operations required to compute the product1. For example, if we have three matrices A, B and C of sizes 10x100, 100x5 and 5x50 respectively, then multiplying them in the order (AxB)xC requires 75,000 scalar operations, while multiplying them in the order Ax(BxC) requires only 12,500 scalar operations.