Create n_queen_code.cpp
The n-queen problem is a puzzle where you have to place n chess queens on an n x n chessboard so that no two queens can attack each other. This means that no two queens can share the same row, column or diagonal. It is a classic example of backtracking, which is a technique of trying different possibilities and undoing them if they don’t work. There are many ways to solve this problem using different algorithms and heuristics.